+++ taichi/analysis/detect_fors_with_break.cpp
@@ -1,6 +1,7 @@
 #include "taichi/ir/ir.h"
-#include "taichi/ir/frontend_ir.h"
+#include "taichi/ir/analysis.h"
 #include "taichi/ir/visitors.h"
+#include "taichi/ir/frontend_ir.h"
 
 #include <unordered_set>
