+++ taichi/analysis/gather_deactivations.cpp
@@ -1,6 +1,6 @@
 #include "taichi/ir/ir.h"
+#include "taichi/ir/analysis.h"
 #include "taichi/ir/visitors.h"
-#include <set>
 
 TLANG_NAMESPACE_BEGIN
