+++ taichi/program/kernel.cpp
@@ -5,6 +5,7 @@
 #include "taichi/program/async_engine.h"
 #include "taichi/codegen/codegen.h"
 #include "taichi/backends/cuda/cuda_driver.h"
+#include "taichi/ir/passes.h"
 
 TLANG_NAMESPACE_BEGIN
