+++ taichi/transforms/constant_fold.cpp
@@ -1,4 +1,5 @@
 #include "taichi/ir/ir.h"
+#include "taichi/ir/passes.h"
 #include "taichi/ir/visitors.h"
 #include "taichi/program/program.h"
 #include <deque>
