+++ taichi/transforms/fix_block_parents.cpp
@@ -1,4 +1,5 @@
 #include "taichi/ir/ir.h"
+#include "taichi/ir/passes.h"
 #include "taichi/ir/visitors.h"
 
 TLANG_NAMESPACE_BEGIN
