+++ taichi/transforms/insert_scratch_pad.cpp
@@ -1,4 +1,6 @@
 #include "taichi/ir/ir.h"
+#include "taichi/ir/passes.h"
+#include "taichi/ir/analysis.h"
 #include "taichi/ir/visitors.h"
 #include "taichi/ir/scratch_pad.h"
