+++ taichi/transforms/simplify.cpp
@@ -1,4 +1,6 @@
 #include "taichi/ir/ir.h"
+#include "taichi/ir/passes.h"
+#include "taichi/ir/analysis.h"
 #include "taichi/ir/visitors.h"
 #include <set>
 #include <unordered_set>
