+++ taichi/transforms/slp_vectorize.cpp
@@ -1,6 +1,7 @@
 // Superword level vectorization
 
 #include "taichi/ir/ir.h"
+#include "taichi/ir/passes.h"
 #include "taichi/ir/visitors.h"
 #include <typeinfo>
 #include <set>
