+++ taichi/transforms/variable_optimization.cpp
@@ -1,4 +1,5 @@
 #include "taichi/ir/ir.h"
+#include "taichi/ir/passes.h"
 #include "taichi/ir/visitors.h"
 #include "taichi/ir/state_machine.h"
 #include <unordered_map>
