+++ taichi/transforms/vector_split.cpp
@@ -1,6 +1,7 @@
 // Split vectors wider than machine vector width into multiple vectors
 
 #include "taichi/ir/ir.h"
+#include "taichi/ir/passes.h"
 #include "taichi/ir/visitors.h"
 
 TLANG_NAMESPACE_BEGIN
