+++ taichi/transforms/whole_kernel_cse.cpp
@@ -1,4 +1,6 @@
 #include "taichi/ir/ir.h"
+#include "taichi/ir/passes.h"
+#include "taichi/ir/analysis.h"
 #include "taichi/ir/visitors.h"
 
 TLANG_NAMESPACE_BEGIN
