+++ tests/cpp/test_simplify.cpp
@@ -1,4 +1,5 @@
 #include "taichi/ir/frontend.h"
+#include "taichi/ir/passes.h"
 #include "taichi/common/testing.h"
 
 TLANG_NAMESPACE_BEGIN
